Fixed wrong import. cursor error, added MYSQL port
The original requirements.txt pointed to https://github.com/zacharyvoase/slugify which is probably not the version that we should use, as it throws errors.

I got an error when using the code as-is, and this was fixed by setting the mysql connection to buffered (I was connecting to a remote server).

My server was not running on port 3306, so added the option to specify the port.